# internal/apiv1/openapi.yaml
openapi: 3.1.0
info:
  title: wpcalc API v1
  version: "1.0.0"
  description: |
    A stateless JSON REST mirror of wpcalc's server-rendered app, for
    programmatic clients. Authenticated with a bearer API token (`wpcalc
    token create <username>` on the CLI), not the browser session cookie —
    there is no login/logout/tenant-switch operation here, because a bearer
    token is not a session: every request names its tenant explicitly in
    the path instead of relying on a server-side "active tenant".

    This document is generated code's source of truth (`internal/apiv1` is
    generated from it via oapi-codegen) and is itself served by the API at
    `GET /api/v1/openapi.json` and `GET /api/v1/openapi.yaml`.

    Authorization mirrors the RBAC96 model exactly: every operation below
    carries `x-required-permission`/`x-permission-scope`, checked the same
    way as the HTML app (`Identity.CanSystemWide`/`CanInTenant`/`Can`) —
    see `internal/httpx/openapi.yaml` (served at `GET /openapi.json` /
    `/openapi.yaml` / `/openapi.html` by the HTML app) for the full RBAC96
    write-up.
  license:
    name: MIT
    url: https://github.com/rolsim/wpcalc/blob/main/LICENSE
servers:
  - url: /api/v1
security:
  - bearerToken: []
paths:
...
  /tenants:
    get:
      operationId: listTenants
      summary: List every tenant
      x-required-permission: manage_tenants
      x-permission-scope: system
      responses:
        "200":
          description: Every tenant.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Tenant" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createTenant
      summary: Create a tenant
      x-required-permission: manage_tenants
      x-permission-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        default: { $ref: "#/components/responses/Error" }
...
  /tenants/{tenantId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: getTenant
      summary: Fetch one tenant
      x-required-permission: manage_tenants
      x-permission-scope: system
      responses:
        "200":
          description: The tenant.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateTenant
      summary: Rename a tenant
      x-required-permission: manage_tenants
      x-permission-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TenantCreate" }
      responses:
        "200":
          description: Renamed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tenant" }
        default: { $ref: "#/components/responses/Error" }
...
  /roles:
    get:
      operationId: listRoles
      summary: The role catalog
      x-required-permission: manage_roles
      x-permission-scope: system
      responses:
        "200":
          description: Every role, with its permissions.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Role" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createRole
      summary: Create a role
      x-required-permission: manage_roles
      x-permission-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleCreate" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Role" }
        default: { $ref: "#/components/responses/Error" }
  /roles/{roleId}:
    parameters:
      - $ref: "#/components/parameters/RoleId"
    delete:
      operationId: deleteRole
      summary: Delete a role
      description: Fails while still referenced by any assignment or permission grant.
      x-required-permission: manage_roles
      x-permission-scope: system
      responses:
        "204":
          description: Deleted.
        default: { $ref: "#/components/responses/Error" }
  /roles/{roleId}/permissions:
    parameters:
      - $ref: "#/components/parameters/RoleId"
    post:
      operationId: addRolePermission
      summary: Grant a role a permission
      x-required-permission: manage_roles
      x-permission-scope: system
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RolePermissionGrant" }
      responses:
        "204":
          description: Granted.
        default: { $ref: "#/components/responses/Error" }
  /roles/{roleId}/permissions/{permissionId}:
    parameters:
      - $ref: "#/components/parameters/RoleId"
      - $ref: "#/components/parameters/PermissionId"
    delete:
      operationId: removeRolePermission
      summary: Revoke a permission from a role
      x-required-permission: manage_roles
      x-permission-scope: system
      responses:
        "204":
          description: Revoked.
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: >-
        An API token minted by `wpcalc token create <username>`. Sent as
        `Authorization: Bearer <token>`. Tokens are opaque, hashed at rest
        (SHA-256), shown in full only once (at creation time), and can be
        revoked individually without touching the account's password or any
        browser session.
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { type: string, pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$" }
...
    RoleId:
      name: roleId
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9_]+$' }
    PermissionId:
      name: permissionId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/PermissionKey" }
...
  responses:
    Error:
      description: An error, at any non-2xx status.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: A short, stable machine-readable error code (not localized).
          examples: ["not_found"]
        message:
          type: string
          description: >-
            Optional diagnostic detail — which field failed validation and
            why, a decode error, and the like. Not stable, not localized,
            not meant to be pattern-matched by a client; `error` is the
            field to branch on.
...
    Tenant:
      type: object
      required: [id, name]
      properties:
        id: { type: string, pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$" }
        name: { type: string }
    TenantCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
...
    PermissionKey:
      type: string
      enum:
        - manage_tenants
        - manage_roles
        - manage_employees
        - manage_users
        - read
        - print
        - write
    Scope:
      type: string
      enum: [system, tenant, employee]
...
    Role:
      type: object
      required: [id, name, scope, permissions]
      properties:
        id: { type: string, pattern: '^[a-z0-9_]+$' }
        name: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        permissions:
          type: array
          items: { $ref: "#/components/schemas/PermissionKey" }
    RoleCreate:
      type: object
      required: [id, name, scope]
      properties:
        id: { type: string, pattern: '^[a-z0-9_]+$' }
        name: { type: string, minLength: 1 }
        scope: { $ref: "#/components/schemas/Scope" }
    RolePermissionGrant:
      type: object
      required: [permissionId]
      properties:
        permissionId: { $ref: "#/components/schemas/PermissionKey" }
...